switching from double to long double

+++ base/include/amount.hpp
@@ -7,8 +7,8 @@ namespace sakurajin{
         namespace base{
             class amount : public unit_t<5>{
             public:
-                amount(double val);
-                amount(double val, double mult);
+                amount(long double val);
+                amount(long double val, long double mult);
                 
             };
