+++ base/include/mass.hpp
@@ -7,8 +7,8 @@ namespace sakurajin{
         namespace base{
             class mass : public unit_t<3>{
             public:
-                mass(double val);
-                mass(double val, double mult);
+                mass(long double val);
+                mass(long double val, long double mult);
                 
             };
