+++ base/src/amount.cpp
@@ -2,5 +2,5 @@
 
 using namespace sakurajin::unit_system::base;
 
-amount::amount(double val): unit_t<5>(val){}
-amount::amount(double val, double mult): unit_t<5>(val,mult){}
+amount::amount(long double val): unit_t<5>(val){}
+amount::amount(long double val, long double mult): unit_t<5>(val,mult){}
