+++ base/src/electric_current.cpp
@@ -2,5 +2,5 @@
 
 using namespace sakurajin::unit_system::base;
 
-electric_current::electric_current(double val): unit_t<6>(val){}
-electric_current::electric_current(double val, double mult): unit_t<6>(val,mult){}
+electric_current::electric_current(long double val): unit_t<6>(val){}
+electric_current::electric_current(long double val, long double mult): unit_t<6>(val,mult){}
