+++ base/src/luminous_intensity.cpp
@@ -2,5 +2,5 @@
 
 using namespace sakurajin::unit_system::base;
 
-luminous_intensity::luminous_intensity(double val): unit_t<7>(val){}
-luminous_intensity::luminous_intensity(double val, double mult): unit_t<7>(val,mult){}
+luminous_intensity::luminous_intensity(long double val): unit_t<7>(val){}
+luminous_intensity::luminous_intensity(long double val, long double mult): unit_t<7>(val,mult){}
