+++ base/src/mass.cpp
@@ -2,5 +2,5 @@
 
 using namespace sakurajin::unit_system::base;
 
-mass::mass(double val): unit_t<3>(val){}
-mass::mass(double val, double mult): unit_t<3>(val,mult){}
+mass::mass(long double val): unit_t<3>(val){}
+mass::mass(long double val, long double mult): unit_t<3>(val,mult){}
