+++ base/src/time.cpp
@@ -2,5 +2,5 @@
 
 using namespace sakurajin::unit_system::base;
 
-time::time(double val): unit_t<1>(val){}
-time::time(double val, double mult): unit_t<1>(val,mult){}
+time::time(long double val): unit_t<1>(val){}
+time::time(long double val, long double mult): unit_t<1>(val,mult){}
